requirements.txt
langchain
langchain-community
langchain-openai
langchain_chroma
pypdf
gradio
python-dotenv
flask